+++ lol_monitor.py
@@ -1,20 +1,19 @@
 #!/usr/bin/env python3
 """
 Author: Michal Szymanski <misiektoja-github@rm-rf.ninja>
-v1.0
+v1.1
 
 Script implementing real-time monitoring of LoL (League of Legends) players activity:
 https://github.com/misiektoja/lol_monitor/
 
 Python pip3 requirements:
 
-cassiopeia
-pytz
+pulsefire
 python-dateutil
 requests
 """
 
-VERSION=1.0
+VERSION=1.1
 
 # ---------------------------
 # CONFIGURATION SECTION START
@@ -34,17 +33,8 @@ LOL_ACTIVE_CHECK_INTERVAL=60 # 1 min
 # Stats are not available right away, so the default 2 mins is a good compromise
 LOL_GAME_FINISHED_CHECK_INTERVAL=120 # 2 min
 
-# Specify your local time zone so we convert RIOT API timestamps to your time
-LOCAL_TIMEZONE='Europe/Warsaw'
-
-# Cassiopeia settings file location
-CASSIOPEIA_SETTINGS_JSON_FILE="cassiopeia_settings.json"
-
 # How often do we perform alive check by printing "alive check" message in the output; in seconds
 TOOL_ALIVE_INTERVAL=21600 # 6 hours
-
-# Default value for network-related timeouts in functions + alarm signal handler; in seconds
-FUNCTION_TIMEOUT=30 # 30 seconds
 
 # URL we check in the beginning to make sure we have internet connectivity
 CHECK_INTERNET_URL='http://www.google.com/'
@@ -67,7 +57,7 @@ SENDER_EMAIL = "your_sender_email"
 #SENDER_EMAIL = "your_sender_email"
 RECEIVER_EMAIL = "your_receiver_email"
 
-# The name of the .log file; the tool by default will output its messages to lol_monitor_username.log file
+# The name of the .log file; the tool by default will output its messages to lol_monitor_riotidname.log file
 lol_logfile="lol_monitor"
 
 # Value used by signal handlers increasing/decreasing the check for player activity when user is in game (LOL_ACTIVE_CHECK_INTERVAL); in seconds
@@ -80,7 +70,26 @@ LOL_ACTIVE_CHECK_SIGNAL_VALUE=30 # 30 seconds
 TOOL_ALIVE_COUNTER=TOOL_ALIVE_INTERVAL/LOL_CHECK_INTERVAL
 
 stdout_bck = None
-csvfieldnames = ['Match Start', 'Match Stop', 'Duration', 'Victory', 'Kills', 'Deaths', 'Assists', 'Champion', 'Blue team', 'Red team']
+csvfieldnames = ['Match Start', 'Match Stop', 'Duration', 'Victory', 'Kills', 'Deaths', 'Assists', 'Champion', 'Team 1', 'Team 2']
+
+regions_short_to_long = {
+            "eun1": "europe",   # Europe Nordic & East
+            "euw1": "europe",   # Europe West
+            "na1": "americas",  # North America
+            "br1": "americas",  # Brazil
+            "la1": "americas",  # Latin America North
+            "la2": "americas",  # Latin America South
+            "jp1": "asia",      # Japan
+            "kr": "asia",       # Korea
+            "tr1": "asia",      # Turkey
+            "ru": "asia",       # Russia
+            "ph2": "sea",      # The Philippines
+            "sg2": "sea",      # Singapore, Malaysia, & Indonesia
+            "tw2": "sea",      # Taiwan, Hong Kong, and Macao
+            "th2": "sea",      # Thailand
+            "vn2": "sea",      # Vietnam
+            "oc1": "sea"        # Oceania
+}
 
 status_notification=False
 
@@ -99,8 +108,8 @@ from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
 import argparse
 import csv
-import pytz
-import cassiopeia
+import asyncio
+from pulsefire.clients import RiotAPIClient
 
 # Logger class to output messages to stdout and log file
 class Logger(object):
@@ -116,14 +125,6 @@ class Logger(object):
 
     def flush(self):
         pass  
-
-# Class used to generate timeout exceptions
-class TimeoutException(Exception):
-    pass
-
-# Signal handler for SIGALRM when the operation times out
-def timeout_handler(sig, frame):
-    raise TimeoutException
 
 # Signal handler when user presses Ctrl+C
 def signal_handler(sig, frame):
@@ -263,25 +264,14 @@ def send_email(subject,body,body_html,use_ssl):
     return 0
 
 # Function to write CSV entry
-def write_csv_entry(csv_file_name, start_date_ts, stop_date_ts, duration_ts, victory, kills, deaths, assists, champion, blue_team, red_team):
+def write_csv_entry(csv_file_name, start_date_ts, stop_date_ts, duration_ts, victory, kills, deaths, assists, champion, team1, team2):
     try:
         csv_file=open(csv_file_name, 'a', newline='', buffering=1)
         csvwriter = csv.DictWriter(csv_file, fieldnames=csvfieldnames, quoting=csv.QUOTE_NONNUMERIC)
-        csvwriter.writerow({'Match Start': start_date_ts, 'Match Stop': stop_date_ts, 'Duration': duration_ts, 'Victory': victory, 'Kills': kills, 'Deaths': deaths, 'Assists': assists, 'Champion': champion, 'Blue team': blue_team, 'Red team': red_team })
+        csvwriter.writerow({'Match Start': start_date_ts, 'Match Stop': stop_date_ts, 'Duration': duration_ts, 'Victory': victory, 'Kills': kills, 'Deaths': deaths, 'Assists': assists, 'Champion': champion, 'Team 1': team1, 'Team 2': team2 })
         csv_file.close()
     except Exception as e:
         raise
-
-# Function to convert UTC string returned by RIOT API to datetime object in specified timezone
-def convert_utc_str_to_tz_datetime(utc_string, timezone):
-    utc_string_sanitize=utc_string.split('+', 1)[0]
-    utc_string_sanitize=utc_string_sanitize.split('.', 1)[0]
-    dt_utc = datetime.strptime(utc_string_sanitize, '%Y-%m-%dT%H:%M:%S')
-
-    old_tz = pytz.timezone("UTC")
-    new_tz = pytz.timezone(timezone)
-    dt_new_tz = old_tz.localize(dt_utc).astimezone(new_tz)
-    return dt_new_tz
 
 # Function to return the timestamp in human readable format; eg. Sun, 21 Apr 2024, 15:08:45
 def get_cur_ts(ts_str=""):
@@ -353,184 +343,307 @@ def decrease_active_check_signal_handler(sig, frame):
     print("* LoL timers: [active check interval: " + display_time(LOL_ACTIVE_CHECK_INTERVAL) + "]")
     print_cur_ts("Timestamp:\t\t")
 
-# Function checking if player is currently in game
-def is_user_in_match(username: str, region: str):
-    cassiopeia.configuration.settings.expire_sinks()
-    cassiopeia.configuration.settings.clear_sinks()
-    summoner = cassiopeia.get_summoner(name=username, region=region)
-    try:
-        match = summoner.current_match
-        if match:
-            return True
-        else:
-            return False
-    except Exception as e:
-        print("is_user_in_match() error:",e)
-        return False
+# Adding new participant to the team (historical matches)
+def add_new_team_member(list_of_teams, teamid, member):
+    if list_of_teams:
+        for team in list_of_teams:
+            if team.get("id")==teamid:
+                team["members"].append(member)
 
-# Function printing details of the current player's match (user is in game)
-def print_current_match(username: str, region: str, last_match_stop_ts: int):
-    summoner = cassiopeia.get_summoner(name=username, region=region)
+# Adding new participant to the team (current match)
+def add_new_current_team_member(list_of_teams, teamid, member):
+    if not list_of_teams:
+        list_of_teams.append({"id": teamid, "members": [ member ]})
+        return
 
-    match = summoner.current_match
+    teamid_exists=False
+    if list_of_teams:
+        for team in list_of_teams:
+            if team.get("id")==teamid:
+                team["members"].append(member)
+                teamid_exists=True
 
-    if match:
+    if not teamid_exists:
+        list_of_teams.append({"id": teamid, "members": [ member ]})
 
-        match_creation=convert_utc_str_to_tz_datetime(str(match.creation),LOCAL_TIMEZONE)
-        match_creation_ts=int(match_creation.timestamp())
-        if match_creation_ts<1000000000:
-            match_creation=datetime.fromtimestamp(int(time.time()))
-            match_creation_str=get_cur_ts("")
-        else:
-            match_creation=datetime.fromtimestamp(int(match_creation_ts))
-            match_creation_str=get_date_from_ts(int(match_creation_ts))
-        match_duration_ts=int(match.duration.total_seconds())
+# Function returning Riot game name & tag line for specified Riot ID
+def get_user_riot_name_tag(riotid: str):
 
-        print("*** LoL user " + username + " is in game now (after " + calculate_timespan(match_creation,int(last_match_stop_ts)) + ")\n")
+    riotid_name=riotid.split('#', 1)[0]
+    riotid_tag=riotid.split('#', 1)[1]
 
-        print("User played last time:\t" + get_date_from_ts(last_match_stop_ts) + "\n")
+    return riotid_name, riotid_tag
 
-        print("Match ID:\t\t" + str(match.id))
+# Function converting Riot ID to PUUID
+async def get_user_puuid(riotid: str, region: str):
 
-        print("\nMatch creation date:\t" + str(match_creation_str))
-        if match_duration_ts > 0:
-            print("Match duration:\t\t" + display_time(int(match.duration.total_seconds())))
-        else:
-            print("Match duration:\t\tjust starting ...")
-            match_duration_ts=0
+    riotid_name, riotid_tag=get_user_riot_name_tag(riotid)
 
-        p = match.participants[summoner]
-        print("\nChampion:\t\t" + str(p.champion.name))
-        print("Map:\t\t\t" + str(match.map.name))
-
-        blue_team=[]
-        print("\nBLUE team:")
-        for mp in match.blue_team.participants:
+    async with RiotAPIClient(default_headers={"X-Riot-Token": RIOT_API_KEY}) as client: 
+            
             try:
-                print("-",str(mp.summoner.name))
-                blue_team.append(mp.summoner.name)
+                account = await client.get_account_v1_by_riot_id(region=regions_short_to_long[region], game_name=riotid_name, tag_line=riotid_tag)
+                puuid=account["puuid"]
             except Exception as e:
-                # Workaround for recent RIOT API changes not reflected in Cassiopeia yet
-                if 'object has already been loaded':
-                    print("-",str(username))
-                    blue_team.append(str(username))
-                else:
-                    print("- (error getting summoner name in match - " + str(e) + ")")
-                    blue_team.append("")
-                continue                
-        print()
+                print("Error while converting Riot ID to PUUID - " + str(e))
+                puuid=0
 
-        red_team=[]
-        print("RED team:")
-        for mp in match.red_team.participants:
+    return puuid
+
+async def get_summoner_details(puuid: str, region: str):
+
+    summoner_id=""
+    summoner_accountid=""
+    summoner_level=""
+
+    async with RiotAPIClient(default_headers={"X-Riot-Token": RIOT_API_KEY}) as client: 
+            
             try:
-                print("-",str(mp.summoner.name))
-                red_team.append(mp.summoner.name)
+                summoner = await client.get_lol_summoner_v4_by_puuid(region=region, puuid=puuid)
+
+                summoner_id=str(summoner["id"])
+                summoner_accountid=str(summoner["accountId"])
+                summoner_level=str(summoner["summonerLevel"])
+
             except Exception as e:
-                # Workaround for recent RIOT API changes not reflected in Cassiopeia yet
-                if 'object has already been loaded':
-                    print("-",str(username))
-                    red_team.append(str(username))
-                else:
-                    print("- (error getting summoner name in match - " + str(e) + ")")
-                    red_team.append("")
-                continue                
-        print()
+                print("Error while getting summoner details - " + str(e))
 
-        return match_creation_ts, match_creation, match_creation_str, blue_team, red_team, match_duration_ts
-    else:
-        print("User is not in game currently")
-        return int(time.time()), datetime.fromtimestamp(int(time.time())), "", [], [], 0
+    return summoner_id, summoner_accountid, summoner_level
 
-# Functioning printing history of matches with relevant details
-def print_match_history(username: str, region: str, matches_num: int, status_notification_flag, csv_file_name):
-    summoner = cassiopeia.get_summoner(name=username, region=region)
-    matches_history = summoner.match_history
+# Functioning returning start & stop timestamps and duration of last played match
+async def get_last_match_tss_duration(puuid: str, region: str):
+    
     match_start_ts = 0
     match_stop_ts = 0
+    match_duration = 0
 
-    for i in reversed(range(matches_num)):
-        match = matches_history[i]
-        match_creation=convert_utc_str_to_tz_datetime(str(match.creation),LOCAL_TIMEZONE)
-        match_start=convert_utc_str_to_tz_datetime(str(match.start),LOCAL_TIMEZONE)
-        match_start_ts=int(match_start.timestamp())
-        match_start=datetime.fromtimestamp(int(match_start_ts))
-        match_stop_ts=match_start_ts+int(match.duration.total_seconds())
-        match_stop=datetime.fromtimestamp(int(match_stop_ts))
-        print("Match ID:\t\t" + str(match.id))
-        print(f"\nMatch start-end date:\t" + get_range_of_dates_from_tss(match_start_ts,match_stop_ts))
-        print("Match duration:\t\t" + display_time(int(match.duration.total_seconds())))
-        last_played=calculate_timespan(int(time.time()), match_stop_ts)       
-        if i==0:
-            print(f"\nUser played last time:\t{last_played} ago")
-
-        p = match.participants[summoner]
-        # Below code to overcome bug with iteration when accessing multiple times
-        u_victory=p.team.win
-        u_champion=p.champion.name
-        u_kills=p.stats.kills
-        u_deaths=p.stats.deaths
-        u_assists=p.stats.assists
-        u_level=p.stats.level
-        print("\nVictory:\t\t" + str(u_victory))
-        print("Champion:\t\t" + str(u_champion))
-        print(f"Kills/Deaths/Assists:\t{u_kills}/{u_deaths}/{u_assists}")
-        print("Level:\t\t\t" + str(u_level))
-        print("Map:\t\t\t" + str(match.map.name))
-
-        blue_team=[]
-        print("\nBLUE team:")
-        for mp in match.blue_team.participants:
-            try:
-                print("-",str(mp.summoner.name))
-                blue_team.append(mp.summoner.name)
-            except Exception as e:
-                # Workaround for recent RIOT API changes not reflected in Cassiopeia yet
-                if 'object has already been loaded':
-                    print("-",str(username))
-                    blue_team.append(str(username))
-                else:
-                    print("- (error getting summoner name in match - " + str(e) + ")")
-                    blue_team.append("")
-                continue
-        print()
-        blue_team_str=" ".join(f"'{x}'" for x in blue_team)
-
-        red_team=[]
-        print("RED team:")
-        for mp in match.red_team.participants:
-            try:
-                print("-",str(mp.summoner.name))
-                red_team.append(mp.summoner.name)
-            except Exception as e:
-                # Workaround for recent RIOT API changes not reflected in Cassiopeia yet
-                if 'object has already been loaded':
-                    print("-",str(username))
-                    red_team.append(str(username))
-                else:
-                    print("- (error getting summoner name in match - " + str(e) + ")")
-                    red_team.append("")
-                continue
-        red_team_str=" ".join(f"'{x}'" for x in red_team)
+    async with RiotAPIClient(default_headers={"X-Riot-Token": RIOT_API_KEY}) as client:
 
         try:
-            if csv_file_name:
-                 write_csv_entry(csv_file_name, match_start, match_stop, display_time(int(match.duration.total_seconds())), u_victory, u_kills, u_deaths, u_assists, u_champion, blue_team_str, red_team_str)
+
+            matches_history = await client.get_lol_match_v5_match_ids_by_puuid(region=regions_short_to_long[region], puuid=puuid, queries={"start": 0, "count": 1})
+
+            match = await client.get_lol_match_v5_match(region=regions_short_to_long[region], id=matches_history[0])
+
+            match_start_ts=int((match["info"]["gameStartTimestamp"])/1000)
+            match_stop_ts=int((match["info"]["gameEndTimestamp"])/1000)
+            match_duration=match["info"]["gameDuration"]
+
         except Exception as e:
-            print("* Cannot write CSV entry -", e)
+            print("Error while getting last match details - " + str(e))
+    
+    return match_start_ts, match_stop_ts, match_duration
 
-        if status_notification_flag and i==0:
-            m_subject="LoL user " + username + " match summary (" + get_range_of_dates_from_tss(match_start_ts,match_stop_ts,short=True) + ", " + display_time(int(match.duration.total_seconds()),granularity=1) + ", " + str(u_victory) + ")"
-            m_body="LoL user " + username + " last match summary\n\nMatch ID: " + str(match.id) + "\n\nMatch start-end date: " + get_range_of_dates_from_tss(match_start_ts,match_stop_ts) + "\nMatch duration: " + display_time(int(match.duration.total_seconds())) + "\n\nVictory: " + str(u_victory) + "\nChampion: " + str(u_champion) + "\nKills/deaths/assists: " + str(u_kills) + "/" + str(u_deaths) + "/" + str(u_assists) + "\nLevel: " + str(u_level) + "\nMap: " + str(match.map.name) + "\n\nBlue team: " + blue_team_str + "\n\nRed team: " + red_team_str + get_cur_ts("\n\nTimestamp: ")
-            print("Sending email notification to",RECEIVER_EMAIL)
-            send_email(m_subject,m_body,"",SMTP_SSL)
+# Function checking if player is currently in game
+async def is_user_in_match(puuid: str, region: str):
 
-        if matches_num>1:
-            print("-----------------------------------------------------------------------------------")
-    return match_start_ts, match_stop_ts
+    async with RiotAPIClient(default_headers={"X-Riot-Token": RIOT_API_KEY}) as client: 
+
+        try:
+            current_match = await client.get_lol_spectator_v5_active_game_by_summoner(region=region, puuid=puuid)
+            if current_match:
+                return True
+        except Exception as e:
+            #print("Error while checking if user is in match - " + str(e))
+            return False
+
+# Function printing details of the current player's match (user is in game)
+async def print_current_match(puuid: str, riotid_name: str, region: str, last_match_start_ts: int, last_match_stop_ts: int):
+
+    async with RiotAPIClient(default_headers={"X-Riot-Token": RIOT_API_KEY}) as client: 
+
+        try:
+            current_match = await client.get_lol_spectator_v5_active_game_by_summoner(region=region, puuid=puuid)
+        except Exception as e:
+            current_match = False
+
+        if current_match:
+
+            match_id=current_match["gameId"]
+            match_start_ts=int((current_match["gameStartTime"])/1000)
+            match_duration=current_match["gameLength"]
+
+            gamemode=current_match["gameMode"]
+
+            if match_start_ts<1000000000:
+                match_start_ts=int(time.time())
+
+            print("*** LoL user " + riotid_name + " is in game now (after " + calculate_timespan(match_start_ts,int(last_match_stop_ts)) + ")\n")
+
+            print("User played last time:\t" + get_range_of_dates_from_tss(last_match_start_ts,last_match_stop_ts) + "\n")
+
+            print("Match ID:\t\t" + str(match_id))
+
+            print("\nMatch start date:\t" + get_date_from_ts(match_start_ts))
+
+            if match_duration > 0:
+                print("Match duration:\t\t" + display_time(int(match_duration)))
+            else:
+                print("Match duration:\t\tjust starting ...")
+                match_duration=0
+
+            current_teams=[]
+
+            for p in current_match["participants"]:
+                u_riotid=p["riotId"]
+                u_riotid_name=u_riotid.split('#', 1)[0]
+                #u_riotid_tag=u_riotid.split('#', 1)[1]
+
+                u_teamid=p["teamId"]
+
+                add_new_current_team_member(current_teams, u_teamid, u_riotid_name)
+
+                if u_riotid_name==riotid_name:
+                    u_champion_id=p["championId"]
+
+                    print(f"\nGame mode:\t\t{gamemode}")
+                    print(f"Champion ID:\t\t{u_champion_id}")               
+
+            current_teams_str=""
+
+            for team in current_teams:
+                teamid_str=f'\nTeam id {team["id"]}:'
+                current_teams_str+=teamid_str + "\n"
+                print(teamid_str)
+
+                for member in team["members"]:
+                    member_str=f"- {member}"
+                    current_teams_str+=member_str + "\n"
+                    print(member_str)
+
+            return match_start_ts, current_teams_str, match_duration
+        else:
+            print("User is not in game currently")
+            return int(time.time()), "", 0
+
+# Functioning printing history of matches with relevant details
+async def print_match_history(puuid: str, riotid_name: str, region: str, matches_min: int, matches_num: int, status_notification_flag, csv_file_name):
+    
+    match_start_ts = 0
+    match_stop_ts = 0
+    match_duration = 0
+
+    async with RiotAPIClient(default_headers={"X-Riot-Token": RIOT_API_KEY}) as client:
+
+        matches_history = await client.get_lol_match_v5_match_ids_by_puuid(region=regions_short_to_long[region], puuid=puuid, queries={"start": 0, "count": matches_num})
+
+        for i in reversed(range(matches_min-1, matches_num)):
+
+            match = await client.get_lol_match_v5_match(region=regions_short_to_long[region], id=matches_history[i])
+
+            match_id=match["metadata"]["matchId"]
+            match_creation_ts=int((match["info"]["gameCreation"])/1000)
+            match_start_ts=int((match["info"]["gameStartTimestamp"])/1000)
+            match_stop_ts=int((match["info"]["gameEndTimestamp"])/1000)
+            match_duration=match["info"]["gameDuration"]
+
+            gamemode=match["info"]["gameMode"]
+
+            if matches_num>1:
+                print("Match number:\t\t" + str(i+1))
+            print("Match ID:\t\t" + str(match_id))
+            print("Game mode:\t\t" + str(gamemode))
+            print(f"\nMatch start-end date:\t" + get_range_of_dates_from_tss(match_start_ts,match_stop_ts))
+            print("Match creation:\t\t" + get_date_from_ts(match_creation_ts))
+            print("Match duration:\t\t" + display_time(int(match_duration)))
+
+            last_played=calculate_timespan(int(time.time()), match_stop_ts)       
+            if i==0:
+                print(f"\nUser played last time:\t{last_played} ago")
+
+            teams=[]
+
+            for t in match["info"]["teams"]:
+                teams.append({"id": t["teamId"], "win": t["win"], "members": []})
+
+            u_teams_number=len(teams)
+
+            for p in match["info"]["participants"]:
+                u_riotid_name=p.get("riotIdGameName",None)
+                if not u_riotid_name: 
+                    u_riotid_name=p.get("summonerName",None) # supporting old matches
+
+                u_teamid=p["teamId"]
+                
+                add_new_team_member(teams, u_teamid, u_riotid_name)
+
+                if u_riotid_name==riotid_name:
+                    u_victory=p["win"]
+                    u_champion=p["championName"]
+                    u_kills=p["kills"]
+                    u_deaths=p["deaths"]
+                    u_assists=p["assists"]
+                    u_level=p["champLevel"]
+         
+                    print(f"\nVictory:\t\t{u_victory}")
+                    print(f"Champion:\t\t{u_champion}")
+                    print(f"Kills/Deaths/Assists:\t{u_kills}/{u_deaths}/{u_assists}")
+                    print(f"Level:\t\t\t{u_level}")
+                    print(f"Teams:\t\t\t{u_teams_number}")                
+
+            # We display all teams in the console and emails
+            teams_str=""
+
+            # We save only first two teams to CSV file
+            team1=[]
+            team2=[]
+
+            x=0
+            for team in teams:
+                x+=1
+                teamid_str=f'\nTeam id {team["id"]}:'
+                teams_str+=teamid_str + "\n"
+                print(teamid_str)
+
+                for member in team["members"]:
+                    member_str=f"- {member}"
+                    teams_str+=member_str + "\n"
+                    print(member_str)
+                    if x==1:
+                        team1.append(member)
+                    elif x==2:
+                        team2.append(member)
+
+            team1_str=" ".join(f"'{x}'" for x in team1)
+            team2_str=" ".join(f"'{x}'" for x in team2)
+
+            try:
+                if csv_file_name:
+                     write_csv_entry(csv_file_name, str(datetime.fromtimestamp(match_start_ts)), str(datetime.fromtimestamp(match_stop_ts)), display_time(int(match_duration)), u_victory, u_kills, u_deaths, u_assists, u_champion, team1_str, team2_str)
+            except Exception as e:
+                print("* Cannot write CSV entry -", e)
+
+            if status_notification_flag and i==0:
+                m_subject="LoL user " + riotid_name + " match summary (" + get_range_of_dates_from_tss(match_start_ts,match_stop_ts,short=True) + ", " + display_time(int(match_duration),granularity=1) + ", " + str(u_victory) + ")"
+                m_body="LoL user " + riotid_name + " last match summary\n\nMatch ID: " + str(match_id) + "\n\nMatch start-end date: " + get_range_of_dates_from_tss(match_start_ts,match_stop_ts) + "\nMatch creation: " + get_date_from_ts(match_creation_ts) + "\nMatch duration: " + display_time(int(match_duration)) + "\n\nVictory: " + str(u_victory) + "\nChampion: " + str(u_champion) + "\nKills/deaths/assists: " + str(u_kills) + "/" + str(u_deaths) + "/" + str(u_assists) + "\nLevel: " + str(u_level) + "\n" + teams_str + get_cur_ts("\nTimestamp: ")
+                print("Sending email notification to",RECEIVER_EMAIL)
+                send_email(m_subject,m_body,"",SMTP_SSL)
+
+            if matches_num>1:
+                print("-----------------------------------------------------------------------------------")
+    
+    return match_start_ts, match_stop_ts, match_duration
+
+# Function printing last n matches for the user
+async def print_save_recent_matches(riotid: str, region: str, matches_min: int, matches_num: int, csv_file_name, csv_exists):
+
+    try:
+        if csv_file_name:
+            csv_file=open(csv_file_name, 'a', newline='', buffering=1)
+            csvwriter = csv.DictWriter(csv_file, fieldnames=csvfieldnames, quoting=csv.QUOTE_NONNUMERIC)
+            if not csv_exists:
+                csvwriter.writeheader()
+            csv_file.close()
+    except Exception as e:
+        print("* Error -", e)
+
+    puuid=await get_user_puuid(riotid, region)
+    riotid_name, riotid_tag=get_user_riot_name_tag(riotid)
+
+    await print_match_history(puuid, riotid_name, region, matches_min, matches_num, False, csv_file_name)
 
 # Main function monitoring gaming activity of the specified LoL user
-def lol_monitor_user(username, region, error_notification, csv_file_name, csv_exists):
+async def lol_monitor_user(riotid, region, error_notification, csv_file_name, csv_exists):
 
     alive_counter = 0
     last_match_start_ts = 0
@@ -547,18 +660,22 @@ def lol_monitor_user(username, region, error_notification, csv_file_name, csv_ex
     except Exception as e:
         print("* Error -", e)
   
-    summoner = cassiopeia.get_summoner(name=username, region=region)
+    puuid=await get_user_puuid(riotid, region)
 
-    print("Summoner name:\t\t" + str(summoner.name))
-    #print("Summoner ID:\t\t" + str(summoner.id))
-    #print("Summoner account ID:\t" + str(summoner.account_id))
-    print("Summoner level:\t\t" + str(summoner.level))
-    print("Summoner region:\t" + str(summoner.region))
+    riotid_name, riotid_tag=get_user_riot_name_tag(riotid)
+
+    summoner_id, summoner_accountid, summoner_level=await get_summoner_details(puuid, region)
+
+    print("RIOT ID (name#tag):\t" + str(riotid))
+    print("PUUID:\t\t\t" + str(puuid))
+    #print("Summoner ID:\t\t" + str(summoner_id))
+    #print("Summoner account ID:\t" + str(summoner_accountid))
+    print("Summoner level:\t\t" + str(summoner_level))
     print()
 
     print("User last played match:\n")
     try:
-        last_match_start_ts,last_match_stop_ts=print_match_history(username, region, 1, False, None)
+        last_match_start_ts,last_match_stop_ts, last_match_duration=await print_match_history(puuid, riotid_name, region, 1, 1, False, None)
     except Exception as e:
         if 'Forbidden' in str(e) or 'Unknown patch name' in str(e):
             print("* API key might not be valid anymore or new patch deployed!")
@@ -573,20 +690,9 @@ def lol_monitor_user(username, region, error_notification, csv_file_name, csv_ex
 
     while True:
 
-        # Sometimes is_user_in_match() can halt, so to overcome this we use alarm signal functionality to kill it inevitably
-        signal.signal(signal.SIGALRM, timeout_handler)
-        signal.alarm(FUNCTION_TIMEOUT)
         try:
-            ingame=is_user_in_match(username, region)
-            signal.alarm(0)
-        except TimeoutException:
-            signal.alarm(0)
-            print("is_user_in_match() timeout, retrying in", display_time(FUNCTION_TIMEOUT))
-            print_cur_ts("Timestamp:\t\t")
-            time.sleep(FUNCTION_TIMEOUT)           
-            continue   
+            ingame=await is_user_in_match(puuid, region)
         except Exception as e:
-            signal.alarm(0)
             print("Retrying in", display_time(LOL_CHECK_INTERVAL), ", error -", e)
             print_cur_ts("Timestamp:\t\t")
             time.sleep(LOL_CHECK_INTERVAL)         
@@ -597,21 +703,19 @@ def lol_monitor_user(username, region, error_notification, csv_file_name, csv_ex
             if ingame != ingameold:
                 if ingame:
 
-                    current_match_creation_ts, current_match_creation, current_match_creation_str, current_blue_team, current_red_team, current_match_duration_ts=print_current_match(username, region, last_match_stop_ts)
+                    current_match_start_ts, current_teams_str, current_match_duration_ts=await print_current_match(puuid, riotid_name, region, last_match_start_ts, last_match_stop_ts)
 
-                    current_blue_team_str=" ".join(f"'{x}'" for x in current_blue_team)
-                    current_red_team_str=" ".join(f"'{x}'" for x in current_red_team)
                     if current_match_duration_ts > 0:
                         current_match_duration=display_time(current_match_duration_ts)
                     else:
                         current_match_duration="just starting ..."
-                    m_subject="LoL user " + username + " is in game now (after " + calculate_timespan(current_match_creation,int(last_match_stop_ts),show_seconds=False) + " - " + get_short_date_from_ts(last_match_stop_ts) + ")"
-                    m_body="LoL user " + username + " is in game now (after " + calculate_timespan(current_match_creation,int(last_match_stop_ts)) + ")\n\nUser played last time: " + get_date_from_ts(last_match_stop_ts) + "\n\nMatch creation date: " + str(current_match_creation_str) + "\n\nMatch duration: " + current_match_duration + "\n\nBlue team: " + current_blue_team_str + "\n\nRed team: " + current_red_team_str + get_cur_ts("\n\nTimestamp: ")
+                    m_subject="LoL user " + riotid_name + " is in game now (after " + calculate_timespan(current_match_start_ts,int(last_match_stop_ts),show_seconds=False) + " - " + get_short_date_from_ts(last_match_stop_ts) + ")"
+                    m_body="LoL user " + riotid_name + " is in game now (after " + calculate_timespan(current_match_start_ts,int(last_match_stop_ts)) + ")\n\nUser played last time: " + get_range_of_dates_from_tss(last_match_start_ts,last_match_stop_ts) + "\n\nMatch start date: " + get_date_from_ts(current_match_start_ts) + "\n\nMatch duration: " + current_match_duration + "\n" + current_teams_str + get_cur_ts("\nTimestamp: ")
                     gamefinished=False
                 else:
-                    print("*** LoL user",username,"stopped playing !")
-                    m_subject="LoL user " + username + " stopped playing"
-                    m_body="LoL user " + username + " stopped playing" + get_cur_ts("\n\nTimestamp: ")
+                    print("*** LoL user",riotid_name,"stopped playing !")
+                    m_subject="LoL user " + riotid_name + " stopped playing"
+                    m_body="LoL user " + riotid_name + " stopped playing" + get_cur_ts("\n\nTimestamp: ")
                     gamefinished=True
 
                 if status_notification:
@@ -626,20 +730,15 @@ def lol_monitor_user(username, region, error_notification, csv_file_name, csv_ex
             # User finished playing the match
             if gamefinished:
                 time.sleep(LOL_GAME_FINISHED_CHECK_INTERVAL)
-                print("*** Getting details of last played match for",username,"...\n")
-                cassiopeia.configuration.settings.expire_sinks()
-                cassiopeia.configuration.settings.clear_sinks()
-                summoner = cassiopeia.get_summoner(name=username, region=region)
-                last_match = summoner.match_history[0]
-                last_match_start=convert_utc_str_to_tz_datetime(str(last_match.start),LOCAL_TIMEZONE)
-                last_match_start_ts=int(last_match_start.timestamp())
-                last_match_start=datetime.fromtimestamp(int(last_match_start_ts))
- 
+                print("*** Getting details of last played match for",riotid_name,"...\n")
+
+                last_match_start_ts, last_match_stop_ts_new, last_match_duration = await get_last_match_tss_duration(puuid, region)
+
                 if last_match_start_ts != last_match_start_ts_old:
-                    last_match_stop_ts=last_match_start_ts+int(last_match.duration.total_seconds())
+                    last_match_stop_ts=last_match_stop_ts_new
 
                     print("User last played match:\n")
-                    last_match_start_ts,last_match_stop_ts=print_match_history(username, region, 1, status_notification, csv_file_name)
+                    last_match_start_ts,last_match_stop_ts, last_match_duration=await print_match_history(puuid, riotid_name, region, 1, 1, status_notification, csv_file_name)
                     last_match_start_ts_old=last_match_start_ts
 
                 gamefinished=False
@@ -652,6 +751,20 @@ def lol_monitor_user(username, region, error_notification, csv_file_name, csv_ex
                 time.sleep(LOL_ACTIVE_CHECK_INTERVAL)
             else:
                 time.sleep(LOL_CHECK_INTERVAL)
+
+                # Below code tries to handle situations when Spectator API is not available, but new matches show up
+                last_match_start_ts_new, last_match_stop_ts_new, last_match_duration_new = await get_last_match_tss_duration(puuid, region)
+
+                if last_match_start_ts_new != last_match_start_ts_old:
+                    last_match_stop_ts=last_match_stop_ts_new
+
+                    print("User last played match:\n")
+                    last_match_start_ts,last_match_stop_ts, last_match_duration=await print_match_history(puuid, riotid_name, region, 1, 1, status_notification, csv_file_name)
+                    last_match_start_ts_old=last_match_start_ts
+
+                    gamefinished=False
+
+                    print_cur_ts("\nTimestamp:\t\t")
 
             if alive_counter >= TOOL_ALIVE_COUNTER:
                 print_cur_ts("Alive check, timestamp: ")
@@ -686,10 +799,11 @@ if __name__ == "__main__":
     print("League of Legends Monitoring Tool",VERSION,"\n")
 
     parser = argparse.ArgumentParser("lol_monitor")
-    parser.add_argument("username", nargs="?", help="User's LoL username", type=str)
+    parser.add_argument("riotid", nargs="?", help="User's LoL Riot ID", type=str)
     parser.add_argument("region", nargs="?", help="User's LoL region", type=str)
     parser.add_argument("-l","--list_recent_matches", help="List recent matches for the user", action='store_true')
-    parser.add_argument("-n", "--number_of_recent_matches", help="Number of recent matches to display if used with -l", type=int)
+    parser.add_argument("-n", "--number_of_recent_matches", help="Number of recent matches to display/save if used with -l and/or -b", type=int)
+    parser.add_argument("-m", "--min_of_recent_matches", help="Minimal match to display/save if used with -l and -n, it will limit range of matches from min_of_recent_matches (e.g. 300) to number_of_recent_matches (e.g. 500)", type=int)    
     parser.add_argument("-b", "--csv_file", help="Write all game playing status changes to CSV file", type=str, metavar="CSV_FILENAME")
     parser.add_argument("-s","--status_notification", help="Send email notification once user changes game playing status", action='store_true')
     parser.add_argument("-e","--error_notification", help="Disable sending email notifications in case of errors like invalid API key", action='store_false')
@@ -699,8 +813,8 @@ if __name__ == "__main__":
     parser.add_argument("-r", "--riot_api_key", help="Specify RIOT API key if not defined within the script", type=str)
     args = parser.parse_args()
 
-    if not args.username or not args.region:
-        print("* username and region arguments are required")
+    if not args.riotid or not args.region:
+        print("* riotid and region arguments are required")
         sys.exit(1)
 
     sys.stdout.write("* Checking internet connectivity ... ")
@@ -718,24 +832,6 @@ if __name__ == "__main__":
     if args.riot_api_key:
         RIOT_API_KEY=args.riot_api_key
 
-    cassiopeia.apply_settings(CASSIOPEIA_SETTINGS_JSON_FILE)
-
-    cassiopeia.set_riot_api_key(RIOT_API_KEY)
-
-    if args.list_recent_matches:
-        if args.number_of_recent_matches and args.number_of_recent_matches>0:
-            matches_n=args.number_of_recent_matches
-        else:
-            matches_n=2
-        print("* Listing " + str(matches_n) + " recent matches for " + str(args.username) + ":\n")
-        try:
-            print_match_history(args.username, args.region, matches_n, False, None)
-        except Exception as e:                     
-            if 'Forbidden' in str(e) or 'Unknown patch name' in str(e):
-                print("* API key might not be valid anymore or new patch deployed!")
-            print("* Error -", e)
-        sys.exit(0)
-
     if args.csv_file:
         csv_enabled=True
         csv_exists=os.path.isfile(args.csv_file)
@@ -750,8 +846,42 @@ if __name__ == "__main__":
         csv_file=None
         csv_exists=False
 
+    if args.list_recent_matches:
+        if args.number_of_recent_matches and args.number_of_recent_matches>0:
+            matches_num=args.number_of_recent_matches
+        else:
+            matches_num=2
+
+        if args.min_of_recent_matches and args.min_of_recent_matches>0:
+            matches_min=args.min_of_recent_matches
+        else:
+            matches_min=1
+
+        if matches_min>matches_num:
+            print("* min_of_recent_matches cannot be > number_of_recent_matches")
+            sys.exit(1)
+
+        if args.csv_file:
+            list_operation="* Listing & saving "
+        else:
+            list_operation="* Listing "
+
+        if matches_min!=matches_num:
+            print(list_operation + "recent matches from " + str(matches_num) + " to " + str(matches_min) + " for " + str(args.riotid) + ":\n")
+        else:
+            print(list_operation + "recent match for " + str(args.riotid) + ":\n")
+
+        try:
+            asyncio.run(print_save_recent_matches(args.riotid, args.region, matches_min, matches_num, args.csv_file,csv_exists))
+        except Exception as e:                     
+            if 'Forbidden' in str(e) or 'Unknown patch name' in str(e):
+                print("* API key might not be valid anymore or new patch deployed!")
+            print("* Error -", e)
+        sys.exit(0)
+
+    riotid_name, riotid_tag=get_user_riot_name_tag(args.riotid)
     if not args.disable_logging:
-        lol_logfile = lol_logfile + "_" + str(args.username) + ".log"
+        lol_logfile = lol_logfile + "_" + str(riotid_name) + ".log"
         sys.stdout = Logger(lol_logfile)
 
     status_notification=args.status_notification
@@ -765,7 +895,7 @@ if __name__ == "__main__":
     signal.signal(signal.SIGTRAP, increase_active_check_signal_handler)
     signal.signal(signal.SIGABRT, decrease_active_check_signal_handler)
 
-    lol_monitor_user(args.username,args.region,args.error_notification,args.csv_file,csv_exists)
+    asyncio.run(lol_monitor_user(args.riotid,args.region,args.error_notification,args.csv_file,csv_exists))
 
     sys.stdout = stdout_bck
     sys.exit(0)